requirements.txt
fastapi[all]
sqlalchemy
uvicorn
pydantic
omegaconf
python_dotenv
psycopg2-binary
pytest
ruff